+++ HISTORY.md
@@ -1,3 +1,10 @@
+0.0.2 / 2016-07-23
+==================
+
+  * Implemented metasync.composition, issue #1
+  * Implemented metasync.parallel
+  * Implemented metasync.sequential
+
 0.0.1 / 2016-07-18
 ==================